// admin/src/api/generated/schemas/adminVaultBalanceDetailItem.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Internal Wallet Admin API
 * Admin API (served via api-gateway)
 * OpenAPI spec version: v1
 */
import type { AdminVaultThresholdDetail } from './adminVaultThresholdDetail';

export type AdminVaultBalanceDetailItem = {
  currency?: string;
  contract_address?: string;
  balance?: string;
  balance_usd?: string;
  locked_balance?: string;
  available_balance?: string;
  pending_withdrawals?: number;
  pending_withdrawals_amount?: string;
  threshold_config?: AdminVaultThresholdDetail;
  status?: string;
};