// admin/src/api/generated/schemas/adminVaultRecentTransactionItem.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Internal Wallet Admin API
 * Admin API (served via api-gateway)
 * OpenAPI spec version: v1
 */

export type AdminVaultRecentTransactionItem = {
  id?: string;
  type?: string;
  currency?: string;
  amount?: string;
  to_address?: string;
  tx_hash?: string;
  status?: string;
  created_at?: string;
};